Move code and fix exports.
